Changed back to double.

+++ include/mcltypes.h
@@ -20,7 +20,7 @@
 #include <vector>
 
 namespace mcl {
-typedef float Real; /**< Real type */
+typedef double Real; /**< Real type */
 typedef std::complex<Real> Complex; /**< Complex type */
 typedef unsigned long UInt; /**< Unisgned int type */
 typedef long Int; /**< Int type */
